Change GCLOUD_TABLE to a input instead of a secret.
GitHub Actions filters any output values that contain a secret.
Turns out, this started matching one of the images, which started causing
some problems since the matrix generation couldn't parse an empty
string.

This value isn't really sensitive, but secrets happened to be a convinient way
to set this out of band of the action itself. Moving to an input so that
this can still be configured, but will default to the current value.